py/requirements.txt
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.0.1
PyAutoGUI==0.9.54
keyboard==0.13.5
requests==2.31.0
python-dotenv==1.0.0
